requirements/requirements.txt
# torch==2.4.0
# torchvision==0.19.0

# # may need to build from source
# apex

# # use our version
# colossalai>=0.4.3 

# # choose the correct wheel packages or build from source
# xformers>=0.0.27
# flash-attn>=2.6.3

# can use pip install -r requirements.txt
mmcv==1.7.2
mmdet==2.28.2
mmengine>=0.10.3
timm==0.9.16
rotary-embedding-torch==0.5.3
diffusers==0.30.0 # for vae
accelerate>=0.29.3 # for t5
ftfy>=6.2.0 # for t5
numpy==1.24.2
numba==0.60.0
tensorboard>=2.14.0
fire
hydra-core==1.3.0
h5py>=3.8.0
moviepy>=1.0.3
nuscenes-devkit==1.1.11
opencv-python==4.10.0.84
